+++ R/module-compliance.R
@@ -14,6 +14,10 @@ compliance_server <- function(id, sessions, sessions_colnames) {
   shiny::moduleServer(id, function(input, output, session) {
     compliance_table <- shiny::reactive({
       shiny::req(sessions())
+      if (is.null(sessions_colnames()$night)) {
+        shiny::HTML("'Night' column was not specified.<br/>Please set a column name for night.")
+      }
+      shiny::validate(shiny::need(!is.null(sessions_colnames()$night), message = FALSE))
       get_compliance_table(sessions(), sessions_colnames())
     })
 
@@ -46,6 +50,9 @@ compliance_server <- function(id, sessions, sessions_colnames) {
 
     shiny::observe({
       shiny::req(sessions())
+      shiny::validate(
+        shiny::need(!is.null(sessions_colnames()$night), message = FALSE)
+      )
       output_table <- sessions() |>
         dplyr::filter(.data$display) |>
         get_non_complying_sessions(col_names = sessions_colnames())
@@ -66,30 +73,35 @@ get_compliance_table <- function(sessions, col_names = NULL) {
     make_sessions_display_table(col_names = col_names)
 }
 
-#' @importFrom rlang .data
 make_sessions_display_table <- function(sessions, col_names = NULL) {
-  col <- get_session_colnames(sessions, col_names)
+  col <- col_names
   sessions <- sessions |>
     dplyr::mutate(
-      start_time = parse_time(.data[[col$session_start]]) |> format("%H:%M"),
-      sleep_onset = parse_time(.data[[col$time_at_sleep_onset]]) |> format("%H:%M"),
-      wakeup_time = parse_time(.data[[col$time_at_wakeup]]) |> format("%H:%M"),
-      end_time = parse_time(.data[[col$session_end]]) |> format("%H:%M"),
-      session_duration_h = difftime(parse_time(.data[[col$session_end]]),
-                                    parse_time(.data[[col$session_start]]),
+      start = parse_time(get_col(sessions, col$session_start)) |> format("%H:%M"),
+      sleep_onset = parse_time(get_col(sessions, col$time_at_sleep_onset)) |> format("%H:%M"),
+      wakeup = parse_time(get_col(sessions, col$time_at_wakeup)) |> format("%H:%M"),
+      end = parse_time(get_col(sessions, col$session_end)) |> format("%H:%M"),
+      session_duration_h = difftime(parse_time(get_col(sessions, col$session_end)),
+                                    parse_time(get_col(sessions, col$session_start)),
                                     units = "hours"),
-      night = lubridate::as_date(.data[[col$night]]) |> format("%Y-%m-%d"),
-      time_in_bed_h = if (!is.null(col$time_in_bed)) .data[[col$time_in_bed]] / 60 / 60 else NA
+      night = if (!is.null(col$night)) {
+        lubridate::as_date(get_col(sessions, col$night)) |> format("%Y-%m-%d")
+      } else {
+        NA
+      },
+      time_in_bed_h = if (!is.null(col$time_in_bed)) get_col(sessions, col$time_in_bed) / 60 / 60 else NA
     ) |>
     dplyr::select(
-      col$id,
-      "night",
-      "start_time",
-      "sleep_onset",
-      "wakeup_time",
-      "end_time",
-      "session_duration_h",
-      "time_in_bed_h",
-      "annotation"
+      dplyr::any_of(c(
+        col$id,
+        "night",
+        "start",
+        "sleep_onset",
+        "wakeup",
+        "end",
+        "session_duration_h",
+        "time_in_bed_h",
+        "annotation"
+      ))
     )
 }